Installer, úprava splitu

+++ bank-server/main.lua
@@ -1,3 +1,5 @@
+-- Bank server
+
 term.clear()
 term.setCursorPos(10,1)
 term.setTextColor(colors.orange)
@@ -5,17 +7,18 @@ print("BANK SERVER")
 term.setCursorPos(1,3)
 rednet.open("top")
 
-function string:split(sep)
-	local sep, fields = sep or ":", {}
-	local pattern = string.format("([^%s]+)", sep)
-	self:gsub(pattern, function(c) fields[#fields+1] = c end)
-	return fields
+function split(message)
+	local args = {}
+	for arg in message:gmatch('%S+') do
+		table.insert(args, arg)
+	end
+	return args
 end
 
 while true do
 	local senderId, message, protocol = rednet.receive("BANK:4471")
 	if message:match(";") then
-		local cmd, arg1, arg2 = unpack(message:split(";"))
+		local cmd, arg1, arg2 = unpack(split(message))
 		arg1 = tonumber(arg1)
 		arg2 = tonumber(arg2)
 		local money = 0
@@ -33,7 +36,7 @@ while true do
 		if(cmd == "check")then
 			rednet.send(senderId, money, "BANK:4471")
 		elseif(cmd == "send")then
-			if arg1 and arg2 and fs.exists("bank/"..arg1) and arg2 <= money then
+			if arg1 and arg2 and fs.exists("bank/"..arg1) and arg2 > 0 and arg2 <= money then
 				local file = fs.open("bank/"..senderId, "w")
 				file.write(money - arg2)
 				file.close()
